requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.2
django-admin-interface==0.22.1
django-colorfield==0.7.2
django-crispy-forms==1.14.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
gunicorn==19.9.0
Pillow==9.2.0
psycopg2==2.9.5
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
whitenoise==6.2.0
